adapt to dkb.de html change (#16)
The transaction selection form used to contain a select
field which allowed for selection of the appropriate
bank account.
This field was looked up by name.  It is still there.
However, there is now a second, hidden field, which has
the same name. This causes the current logic to break.
The workaround consists of looking explicitly for a
'select' type field.

Fixes #16.